+++ CataVentoApi/Entity/Group.cs
@@ -2,7 +2,7 @@
 {
     public class Group
     {
-        public long GroupId { get; set; }
+        public int GroupId { get; set; }
         public string GroupName { get; set; }
         public List<long> UsuariosIds { get; set; } = new List<long>();
     }
